feat!: moves general model handling and utilities to sdk pkg
