+++ tests/trestle/core/commands/author/component_test.py
@@ -25,7 +25,7 @@ import trestle.sdk.models.generic_oscal as generic
 import trestle.sdk.oscal.component as comp
 from trestle.common import const, file_utils, model_utils
 from trestle.core.commands.common.return_codes import CmdReturnCodes
-from trestle.core.control_interface import ControlInterface
+from trestle.sdk.models.control_interface import ControlInterface
 from trestle.core.markdown.markdown_processor import MarkdownProcessor
 
 md_path = 'md_comp'
