+++ tests/trestle/core/crm/export_writer_test.py
@@ -23,7 +23,7 @@ import pytest
 
 from tests import test_utils
 
-import trestle.core.generators as gens
+import trestle.sdk.models.generators as gens
 import trestle.sdk.oscal.ssp as ossp
 from trestle.common.err import TrestleError
 from trestle.common.model_utils import ModelUtils
