+++ tests/trestle/core/repository_test.py
@@ -21,11 +21,12 @@ import pytest
 from tests import test_utils
 
 import trestle.common.const as const
+from trestle.sdk.models import generators
 import trestle.sdk.oscal as oscal
 import trestle.sdk.oscal.catalog as cat
 import trestle.sdk.oscal.profile as prof
 from trestle.common.err import TrestleError
-from trestle.core import generators, parser
+from trestle.core import parser
 from trestle.core.repository import AgileAuthoring, ManagedOSCAL, Repository
 
 prof_name = 'comp_prof'
