+++ trestle/core/catalog/catalog_reader.py
@@ -18,17 +18,17 @@ import pathlib
 from typing import Any, Dict, List, Optional, Tuple, Union
 
 import trestle.common.const as const
-import trestle.core.generators as gens
+import trestle.sdk.models.generators as gens
 import trestle.sdk.models.generic_oscal as generic
 import trestle.sdk.oscal.catalog as cat
 import trestle.sdk.oscal.common as com
 import trestle.sdk.oscal.component as comp
-from trestle.common.common_types import TypeWithSetParams
+from trestle.sdk.models.common_types import TypeWithSetParams
 from trestle.common.err import TrestleError
 from trestle.common.list_utils import as_list, none_if_empty
 from trestle.core.catalog.catalog_interface import CatalogInterface
 from trestle.core.control_context import ControlContext
-from trestle.core.control_interface import CompDict, ComponentImpInfo, ControlInterface
+from trestle.sdk.models.control_interface import CompDict, ComponentImpInfo, ControlInterface
 from trestle.core.control_reader import ControlReader
 from trestle.sdk.oscal import profile as prof
 from trestle.sdk.oscal import ssp as ossp
