+++ trestle/core/commands/split.py
@@ -31,7 +31,7 @@ from trestle.core.models.actions import Action, CreatePathAction, WriteFileActio
 from trestle.core.models.elements import Element, ElementPath
 from trestle.core.models.file_content_type import FileContentType
 from trestle.core.models.plans import Plan
-from trestle.core.trestle_base_model import TrestleBaseModel
+from trestle.sdk.models.trestle_base_model import TrestleBaseModel
 from trestle.sdk.models.base_model import OscalBaseModel
 
 logger = logging.getLogger(__name__)
