+++ trestle/core/models/elements.py
@@ -23,13 +23,14 @@ from pydantic.v1.error_wrappers import ValidationError
 from ruamel.yaml import YAML
 
 import trestle.common.const as const
-from trestle.common import common_types, str_utils
+from trestle.common import str_utils
 from trestle.common import type_utils as utils
 from trestle.common.err import TrestleError, TrestleNotFoundError
 from trestle.common.model_utils import ModelUtils
 from trestle.common.str_utils import AliasMode, classname_to_alias
 from trestle.core.models.file_content_type import FileContentType
 from trestle.sdk.models.base_model import OscalBaseModel
+from trestle.sdk.models import common_types
 
 logger = logging.getLogger(__name__)
