+++ trestle/core/resolver/_import.py
@@ -22,7 +22,7 @@ import trestle.common.const as const
 import trestle.sdk.oscal.catalog as cat
 import trestle.sdk.oscal.profile as prof
 from trestle.common.err import TrestleError
-from trestle.core.control_interface import ParameterRep
+from trestle.sdk.models.control_interface import ParameterRep
 from trestle.core.pipeline import Pipeline
 from trestle.core.remote import cache
 from trestle.core.resolver.merge import Merge
