+++ trestle/core/resolver/merge.py
@@ -19,7 +19,7 @@ from typing import Iterator, List, Optional, Tuple, Union
 import trestle.sdk.oscal.catalog as cat
 import trestle.sdk.oscal.common as com
 import trestle.sdk.oscal.profile as prof
-from trestle.common.common_types import OBT
+from trestle.sdk.models.common_types import OBT
 from trestle.common.err import TrestleError
 from trestle.common.list_utils import as_list, none_if_empty
 from trestle.core.pipeline import Pipeline
