+++ trestle/sdk/oscal/assessment_results.py
@@ -31,15 +31,10 @@ from typing import Any, Dict, List, Optional, Union
 
 from pydantic.v1 import AnyUrl, EmailStr, Extra, Field, conint, constr, validator
 
-from trestle.sdk.models.base_model import OscalBaseModel
-from trestle.sdk.oscal import OSCAL_VERSION_REGEX, OSCAL_VERSION
 import trestle.sdk.oscal.common as common
-from trestle.sdk.oscal.common import AssessmentAssets
-from trestle.sdk.oscal.common import Observation
-from trestle.sdk.oscal.common import RelatedObservation
-from trestle.sdk.oscal.common import SystemComponent
-from trestle.sdk.oscal.common import TaskValidValues
-from trestle.sdk.oscal.common import TokenDatatype
+from trestle.sdk.models.base_model import OscalBaseModel
+from trestle.sdk.oscal import OSCAL_VERSION, OSCAL_VERSION_REGEX
+from trestle.sdk.oscal.common import AssessmentAssets, Observation, RelatedObservation, SystemComponent, TaskValidValues, TokenDatatype
 
 
 class LocalDefinitions1(OscalBaseModel):
